requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
black==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
datasets==3.2.0
dill==0.3.8
distro==1.9.0
evaluate==0.4.3
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0rc1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.26.0
networkx==3.3
nltk==3.9.1
numpy==2.2.2
openai==1.61.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20240706
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.25.3
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.42.0
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
xxhash==3.5.0
yarl==1.18.3
